Humidity and Temperature

+++ vrs_exercise_sensors/Core/Inc/i2c.h
@@ -40,7 +40,7 @@ void MX_I2C1_Init(void);
 
 /* USER CODE BEGIN Prototypes */
 uint32_t i2c_read(uint8_t slave_address, uint8_t register_address, uint8_t number_of_registers);
-uint8_t i2c_write(uint8_t slave_address, uint8_t register_address, uint8_t data, uint8_t number_of_registers);
+uint32_t i2c_write(uint8_t slave_address, uint8_t register_address, uint8_t data, uint8_t number_of_registers);
 /* USER CODE END Prototypes */
 
 #ifdef __cplusplus
